+++ packages/core-mobile/app/new/components/totp/AuthenticatorSetup.tsx
