--- packages/core-mobile/app/new/routes/(authenticator)/settingsTotp/_layout.tsx
@@ -1,16 +1,0 @@
-// import React from 'react'
-// import { Stack } from 'new/components/navigation/Stack'
-// import { SettingsTotpProvider } from 'new/contexts/SettingsTotpProvider'
-
-// export default function SettingsTotpLayout(): JSX.Element {
-//   return (
-//     <SettingsTotpProvider>
-//     <Stack>
-//       <Stack.Screen name="setup" />
-//       <Stack.Screen name="scanQrCode" />
-//       <Stack.Screen name="copyCode" />
-//       <Stack.Screen name="verifyCode" />
-//     </Stack>
-//     </SettingsTotpProvider>
-//   )
-// }
