--- packages/core-mobile/app/new/routes/(authenticator)/totp/scanQrCode.tsx
@@ -1,16 +1,0 @@
-import React from 'react'
-import { useSignupContext } from 'new/contexts/SignupProvider'
-import { ScanQrCode as ScanQrCodeComponent } from '../components/ScanQrCode'
-
-export default function ScanQrCode(): JSX.Element {
-  const { goToVerifyCode, goToEnterCodeManually, totpChallenge } =
-    useSignupContext()
-
-  return (
-    <ScanQrCodeComponent
-      totpChallenge={totpChallenge}
-      onEnterCodeManually={goToEnterCodeManually}
-      onVerifyCode={goToVerifyCode}
-    />
-  )
-}
