+++ packages/core-mobile/app/new/routes/(totp)/verifyCode.tsx
@@ -1,6 +1,6 @@
 import React from 'react'
 import { useSignupContext } from 'new/contexts/SignupProvider'
-import { VerifyCode as VerifyCodeComponent } from '../components/VerifyCode'
+import { VerifyCode as VerifyCodeComponent } from '../../components/totp/VerifyCode'
 
 export default function VerifyCode(): JSX.Element {
   const { onVerifyCode, onVerifySuccess } = useSignupContext()
